Исправление ошибок после тестирования.

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,11 +1,14 @@
 package ru.yandex.practicum.filmorate.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -14,6 +17,7 @@ public class FilmService {
     private final FilmStorage filmStorage;
     private final UserService userService;
 
+    @Autowired
     public FilmService(FilmStorage filmStorage, UserService userService) {
         this.filmStorage = filmStorage;
         this.userService = userService;
@@ -24,20 +28,54 @@ public class FilmService {
     }
 
     public Film create(Film film) {
+        if ((film.getName().isEmpty()) || (film.getDescription().length() > 200)
+                || (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) || (film.getDuration() < 0)) {
+            throw new ValidationException("не выполнены условия: название не может быть пустым;\n" +
+                    "    максимальная длина описания — 200 символов;\n" +
+                    "    дата релиза — не раньше 28 декабря 1895 года;\n" +
+                    "    продолжительность фильма должна быть положительной");
+        }
         return filmStorage.create(film);
     }
 
     public Film updateFilm(Film film) {
-        return filmStorage.updateFilm(film);
+        if ((!film.getName().isEmpty()) && (film.getDescription().length() < 200)
+                && (film.getReleaseDate().isAfter(LocalDate.of(1895, 12, 28)))
+                && (film.getDuration() > 0)) {
+            Optional<Film> optionalFilm = findAll().stream().filter(film1 -> film1.getId() == film.getId()).findFirst();
+            if (optionalFilm.isEmpty()) {
+                throw new NotFoundException("фильма с таким Id не существует");
+            }
+            Film filmUpdate = optionalFilm.get();
+            filmUpdate.setName(film.getName());
+            filmUpdate.setDescription(film.getDescription());
+            filmUpdate.setReleaseDate(film.getReleaseDate());
+            filmUpdate.setDuration(film.getDuration());
+            filmUpdate.setLikes(film.getLikes());
+            return filmStorage.updateFilm(film);
+        } else {
+            throw new ValidationException("не выполнены условия: название не может быть пустым;\n" +
+                    "    максимальная длина описания — 200 символов;\n" +
+                    "    дата релиза — не раньше 28 декабря 1895 года;\n" +
+                    "    продолжительность фильма должна быть положительной");
+        }
     }
 
     public void deleteFilm(int filmId) {
-        Film film = filmStorage.getById(filmId);
+        Optional<Film> optionalFilm = filmStorage.getById(filmId);
+        if (optionalFilm.isEmpty()) {
+            throw new NotFoundException("Фильм не найден.");
+        }
+        Film film = optionalFilm.get();
         filmStorage.deleteFilm(film);
     }
     public Set<Integer> addLikes(int userId, int filmId) {
         User user = userService.getById(userId);
-        Film film = filmStorage.getById(filmId);
+        Optional<Film> optionalFilm = filmStorage.getById(filmId);
+        if ((optionalFilm.isEmpty()) || (user == null)) {
+            throw new NotFoundException("Объект не найден.");
+        }
+        Film film = optionalFilm.get();
         Set<Integer> likes = film.getLikes();
         if (likes.contains(user.getId())) {
             return likes;
@@ -47,8 +85,12 @@ public class FilmService {
     }
 
     public void deleteLikes(int userId, int filmId) {
+        Optional<Film> optionalFilm = filmStorage.getById(filmId);
         User user = userService.getById(userId);
-        Film film = filmStorage.getById(filmId);
+        if ((optionalFilm.isEmpty()) || (user == null)) {
+            throw new NotFoundException("Объект не найден.");
+        }
+        Film film = optionalFilm.get();
         Set<Integer> likes = film.getLikes();
         if (likes.contains(user.getId())) {
             likes.remove(user.getId());
